chore(code-garden): split tasks-api routes/tasks.ts into focused helper modules [2026-W26]
Non-functional refactor from repo audit 2026-W26 (finding
"routes/tasks.ts is 517 lines mixing validation, query building,
sorting, and mapping"). Extracts the pure helpers and DB-touching
siblings into focused modules under services/tasks-api/src/routes/tasks/:

  _constants.ts    -- enum sets, limits, task-type prefixes, dependencyInclude
  _validation.ts   -- parseDate, parseLimit, normalizeString/Tags/DependsOnIds,
                      acceptanceCriteriaText
  _pagination.ts   -- encodeCursor, decodeCursor
  _mapper.ts       -- mapTask, formatTaskTitle, mapTaskComment (+ private mapComment)
  _spec.ts         -- canonicalize, specChecksumForDescription,
                      specDriftMessage, rejectSpecDrift
  _deps.ts         -- connectTags, validateDependsOnIds

tasks.ts shrinks from 711 to 479 lines and now contains only the
Express route handlers and the wiring between them. All 33 vitest
cases (pagination cursor/sort, read endpoints, health, prisma
validate) continue to pass against main without modification.
