+++ docs/repo-audits/2026-W26.md
@@ -168,7 +168,7 @@ AI agents (agents/)
 - `apps/tasks/src/tasksApi.ts:21` — `status: string;  // open | triage | ready | doing | acceptance | done` lists `triage` as valid.
 - No `triage` value exists in the API or Prisma schema.
 
-**[Medium] `routes/tasks.ts` is 517 lines mixing validation, query building, sorting, and mapping** *(persists)*
+**[Medium] `routes/tasks.ts` is 517 lines mixing validation, query building, sorting, and mapping** *(persists)* ✅ [PR #TBD](https://github.com/Stoffer-Industries/sindustries/pull/TBD)
 
 - Targeted unit testing of cursor/sort logic and tag management is harder than it needs to be; any structural change touches the same large file.
