+++ services/tasks-api/src/routes/tasks.ts
@@ -1,259 +1,27 @@
 import { Router } from 'express';
-import { createHash } from 'node:crypto';
-import { prisma } from '../lib/prisma';
-import { badRequest, notFound, sendError } from '../lib/http';
+import { prisma } from '../lib/prisma.ts';
+import { badRequest, notFound, sendError } from '../lib/http.ts';
 
-const DEFAULT_LIMIT = 100;
-const MAX_LIMIT = 10000;
-
-const validStatuses = new Set(['open', 'ready', 'doing', 'acceptance', 'done']);
-const validPriorities = new Set(['low', 'medium', 'high', 'urgent']);
-const validAssignees = new Set(['Tom', 'Quinn', 'Rowan', 'Lox', 'Ivy']);
-const validSorts = new Set(['priority', 'createdAt', 'updatedAt', 'dueAt', 'statusChangedAt']);
-const uuidPattern = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
-const validTaskTypes = new Set(['content', 'code', 'research', 'feature']);
-const taskTypeTitlePrefixes = {
-  content: '✍️',
-  code: '💻',
-  research: '🔎',
-  feature: '🔧'
-};
-
-const dependencyInclude = {
-  dependencies: {
-    include: {
-      dependsOn: {
-        select: {
-          id: true,
-          title: true,
-          status: true,
-          completedAt: true
-        }
-      }
-    }
-  }
-};
-
-const priorityOrder = {
-  urgent: 0,
-  high: 1,
-  medium: 2,
-  low: 3
-};
-
-function parseDate(value) {
-  if (!value) return null;
-  const date = new Date(value);
-  return Number.isNaN(date.valueOf()) ? null : date;
-}
-
-function parseLimit(value) {
-  const n = Number.parseInt(value ?? `${DEFAULT_LIMIT}`, 10);
-  if (Number.isNaN(n) || n <= 0) return DEFAULT_LIMIT;
-  return Math.min(n, MAX_LIMIT);
-}
-
-function encodeCursor(createdAt, id) {
-  return Buffer.from(`${createdAt.toISOString()}::${id}`, 'utf8').toString('base64url');
-}
-
-function decodeCursor(cursor) {
-  if (!cursor) return null;
-
-  try {
-    const decoded = Buffer.from(cursor, 'base64url').toString('utf8');
-    const [createdAtRaw, id] = decoded.split('::');
-    const createdAt = new Date(createdAtRaw);
-
-    if (!id || Number.isNaN(createdAt.valueOf())) {
-      return null;
-    }
-
-    return { createdAt, id };
-  } catch {
-    return null;
-  }
-}
-
-function mapComment(comment) {
-  return {
-    id: comment.id,
-    author: comment.author,
-    text: comment.body,
-    createdAt: comment.createdAt,
-    updatedAt: comment.updatedAt
-  };
-}
-
-function mapTask(task) {
-  const dependsOn = task.dependencies
-    ?.map((dependency) => dependency.dependsOn)
-    .filter(Boolean)
-    .map((dependency) => ({
-      id: dependency.id,
-      title: dependency.title,
-      status: dependency.status,
-      completedAt: dependency.completedAt
-    })) ?? [];
-
-  return {
-    id: task.id,
-    title: formatTaskTitle(task.title, task.taskType),
-    description: task.description,
-    status: task.status,
-    statusChangedAt: task.statusChangedAt,
-    priority: task.priority,
-    dueAt: task.dueAt,
-    completedAt: task.completedAt,
-    assignee: task.assignee,
-    archivedAt: task.archivedAt,
-    blocked: task.blocked ?? false,
-    createdAt: task.createdAt,
-    updatedAt: task.updatedAt,
-    taskType: task.taskType ?? null,
-    specChecksum: task.specChecksum ?? null,
-    tags: task.tags?.map((taskTag) => taskTag.tag?.name).filter(Boolean) ?? [],
-    comments: task.comments?.map(mapComment) ?? [],
-    dependsOn,
-    dependsOnIds: dependsOn.map((dependency) => dependency.id),
-    dependencyBlocked: dependsOn.some((dependency) => dependency.status !== 'done')
-  };
-}
-
-function formatTaskTitle(title, taskType) {
-  const prefix = taskTypeTitlePrefixes[taskType];
-  if (!prefix || typeof title !== 'string') return title;
-
-  const knownPrefixPattern = Object.values(taskTypeTitlePrefixes)
-    .map((value) => value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'))
-    .join('|');
-  const titleWithoutKnownPrefix = title.replace(new RegExp(`^(?:${knownPrefixPattern})\\s+`), '');
-  return `${prefix} ${titleWithoutKnownPrefix}`;
-}
-
-function mapTaskComment(comment) {
-  return mapComment(comment);
-}
-
-function normalizeString(value) {
-  return typeof value === 'string' ? value.trim() : value;
-}
-
-function normalizeTags(tags) {
-  if (!Array.isArray(tags)) return null;
-  const normalized = tags
-    .map((tag) => normalizeString(tag))
-    .filter(Boolean)
-    .map((tag) => tag.toLowerCase());
-
-  return [...new Set(normalized)];
-}
-
-function normalizeDependsOnIds(value) {
-  if (!Array.isArray(value)) return null;
-  const normalized = [];
-  const seen = new Set();
-
-  for (const id of value) {
-    if (typeof id !== 'string' || !uuidPattern.test(id)) return null;
-    if (!seen.has(id)) {
-      seen.add(id);
-      normalized.push(id);
-    }
-  }
-
-  return normalized;
-}
-
-function acceptanceCriteriaText(description) {
-  if (!description) return [];
-  const criteria = [];
-  const re = /^\s*-\s*\[[ xX]\]\s+(.+)$/gm;
-  let match;
-  while ((match = re.exec(description)) !== null) {
-    criteria.push(match[1].trim());
-  }
-  return criteria;
-}
-
-function canonicalize(value) {
-  if (Array.isArray(value)) return value.map(canonicalize);
-  if (value && typeof value === 'object') {
-    return Object.fromEntries(Object.keys(value).sort().map((key) => [key, canonicalize(value[key])]));
-  }
-  return value;
-}
-
-function specChecksumForDescription(description) {
-  const canonical = canonicalize({ acceptanceCriteria: acceptanceCriteriaText(description) });
-  return createHash('sha256').update(JSON.stringify(canonical)).digest('hex');
-}
-
-function specDriftMessage(taskId, stored, current) {
-  return `ACs modified after spec approval -- write a new spec to change scope. Task ${taskId} stored specChecksum \`${stored}\` but current AC checksum is \`${current}\`.`;
-}
-
-function rejectSpecDrift(res, task, description) {
-  if (!task.specChecksum) return false;
-  const current = specChecksumForDescription(description ?? task.description);
-  if (current === task.specChecksum) return false;
-  sendError(res, 409, 'SPEC_CHECKSUM_MISMATCH', specDriftMessage(task.id, task.specChecksum, current));
-  return true;
-}
-
-async function connectTags(tagNames) {
-  if (!tagNames?.length) return [];
-
-  return Promise.all(
-    tagNames.map((name) =>
-      prisma.tag.upsert({
-        where: { name },
-        create: { name },
-        update: {}
-      })
-    )
-  );
-}
-
-async function validateDependsOnIds(res, taskId, dependsOnIds) {
-  if (dependsOnIds.includes(taskId)) {
-    badRequest(res, 'SELF_DEPENDENCY_NOT_ALLOWED', 'Task cannot depend on itself');
-    return false;
-  }
-
-  if (dependsOnIds.length === 0) return true;
-
-  const dependencyTasks = await prisma.task.findMany({
-    where: { id: { in: dependsOnIds } },
-    select: { id: true, archivedAt: true }
-  });
-  const foundIds = new Set(dependencyTasks.map((task) => task.id));
-  const missingId = dependsOnIds.find((id) => !foundIds.has(id));
-  if (missingId) {
-    badRequest(res, 'DEPENDENCY_TASK_NOT_FOUND', `Dependency task not found: ${missingId}`);
-    return false;
-  }
-
-  const archivedTask = dependencyTasks.find((task) => task.archivedAt);
-  if (archivedTask) {
-    badRequest(res, 'ARCHIVED_DEPENDENCY_NOT_ALLOWED', `Dependency task is archived: ${archivedTask.id}`);
-    return false;
-  }
-
-  const directCycle = await prisma.taskDependency.findFirst({
-    where: {
-      taskId: { in: dependsOnIds },
-      dependsOnId: taskId
-    },
-    select: { taskId: true }
-  });
-  if (directCycle) {
-    badRequest(res, 'CIRCULAR_DEPENDENCY_NOT_ALLOWED', 'Direct circular dependencies are not allowed');
-    return false;
-  }
-
-  return true;
-}
+import {
+  dependencyInclude,
+  priorityOrder,
+  validAssignees,
+  validPriorities,
+  validSorts,
+  validStatuses,
+  validTaskTypes
+} from './tasks/_constants.ts';
+import {
+  normalizeDependsOnIds,
+  normalizeString,
+  normalizeTags,
+  parseDate,
+  parseLimit
+} from './tasks/_validation.ts';
+import { decodeCursor, encodeCursor } from './tasks/_pagination.ts';
+import { formatTaskTitle, mapTask, mapTaskComment } from './tasks/_mapper.ts';
+import { rejectSpecDrift } from './tasks/_spec.ts';
+import { connectTags, validateDependsOnIds } from './tasks/_deps.ts';
 
 export const tasksRouter = Router();
